Version support (#12)
Version support

the JSON schema will be stored in a block blob. The name of the blob will contain the information about the version of the schema.
blobName: .schema.v1.json
.schema.v2.json
the constructor of the data container has been modified in order to receive the schema version along with schema object;
the newly blobs will be validated against the schema defined at construct time of the data container;
the schema version information that is used to validate the content of a blob, is saved with the content of the blob.
{
content: '...',	// the stringify json
version: 1 // schema version
}
if you try to modify an existing blob:
the schema version is read from the blob
the content will be validated against the schema with that version.
a data container can have multiple version of schema: the schemas will be loaded and cached on demand (only when you try to modify a blob that has been created with an older version of schema)
for append blobs: the newly appended json objects will follow the schema defined at construct time of the data container.